+++ bin/runtime/src/chain_extension/tests/resources/xor.public_input.bytes
