+++ bin/runtime/src/chain_extension/tests/resources/xor.vk.bytes
